+++ src/main/kotlin/joshxviii/plantz/PazEntities.kt
@@ -49,25 +49,19 @@ object PazEntities {
         }
     }
 
-    private val typeToSunCost = mutableMapOf<EntityType<*>, Int>()
-    fun getSunCostFromType(type: EntityType<*>?) : Int = typeToSunCost[type]?: 0
-
     // region Plants
     @JvmField val SUNFLOWER: EntityType<Sunflower> = registerPlant(
         "sunflower",
         EntityType.Builder.of(::Sunflower, MobCategory.CREATURE),
-        sunCost = 5,
         height = 1.1f,
     )
     @JvmField val PEA_SHOOTER: EntityType<PeaShooter> = registerPlant(
         "peashooter",
         EntityType.Builder.of(::PeaShooter, MobCategory.CREATURE),
-        sunCost = 5,
     )
     @JvmField val WALL_NUT: EntityType<WallNut> = registerPlant(
         "wallnut",
         EntityType.Builder.of(::WallNut, MobCategory.CREATURE),
-        sunCost = 5,
         width = 1.0f,
         height = 1.1f,
         attributes = Plant.Companion.PlantAttributes(
@@ -77,7 +71,6 @@ object PazEntities {
     @JvmField val CHOMPER: EntityType<Chomper> = registerPlant(
         "chomper",
         EntityType.Builder.of(::Chomper, MobCategory.CREATURE),
-        sunCost = 7,
         height = 1.5f,
         attributes = Plant.Companion.PlantAttributes(
             maxHealth = 35.0,
@@ -89,7 +82,6 @@ object PazEntities {
     @JvmField val CHERRY_BOMB: EntityType<CherryBomb> = registerPlant(
         "cherrybomb",
         EntityType.Builder.of(::CherryBomb, MobCategory.CREATURE),
-        sunCost = 10,
         width = 0.625f,
         height = 0.7f,
         attributes = Plant.Companion.PlantAttributes(
@@ -100,7 +92,6 @@ object PazEntities {
     @JvmField val POTATO_MINE: EntityType<PotatoMine> = registerPlant(
         "potatomine",
         EntityType.Builder.of(::PotatoMine, MobCategory.CREATURE),
-        sunCost = 3,
         width = 0.65f,
         height = 0.35f,
         attributes = Plant.Companion.PlantAttributes(
@@ -110,7 +101,6 @@ object PazEntities {
     @JvmField val ICE_PEA_SHOOTER: EntityType<IcePeaShooter> = registerPlant(
         "ice_peashooter",
         EntityType.Builder.of(::IcePeaShooter, MobCategory.CREATURE),
-        sunCost = 7,
         attributes = Plant.Companion.PlantAttributes(
             attackDamage = 2.5,
         )
@@ -118,12 +108,10 @@ object PazEntities {
     @JvmField val REPEATER: EntityType<Repeater> = registerPlant(
         "repeater",
         EntityType.Builder.of(::Repeater, MobCategory.CREATURE),
-        sunCost = 7,
     )
     @JvmField val FIRE_PEA_SHOOTER: EntityType<FirePeaShooter> = registerPlant(
         "fire_peashooter",
         EntityType.Builder.of(::FirePeaShooter, MobCategory.CREATURE).fireImmune(),
-        sunCost = 7,
         attributes = Plant.Companion.PlantAttributes(
             attackDamage = 2.5,
         )
@@ -131,7 +119,6 @@ object PazEntities {
     @JvmField val CACTUS: EntityType<Cactus> = registerPlant(
         "cactus",
         EntityType.Builder.of(::Cactus, MobCategory.CREATURE),
-        sunCost = 6,
         width = 0.8f,
         height = 1.25f,
         eyeHeight = 0.85f,
@@ -143,7 +130,6 @@ object PazEntities {
     @JvmField val CABBAGE_PULT: EntityType<CabbagePult> = registerPlant(
         "cabbagepult",
         EntityType.Builder.of(::CabbagePult, MobCategory.CREATURE),
-        sunCost = 5,
         width = 0.9f,
         height = 0.8f,
         attributes = Plant.Companion.PlantAttributes(
@@ -155,7 +141,6 @@ object PazEntities {
     @JvmField val KERNEL_PULT: EntityType<KernelPult> = registerPlant(
         "kernelpult",
         EntityType.Builder.of(::KernelPult, MobCategory.CREATURE),
-        sunCost = 7,
         width = 0.9f,
         height = 0.8f,
         attributes = Plant.Companion.PlantAttributes(
@@ -167,7 +152,6 @@ object PazEntities {
     @JvmField val MELON_PULT: EntityType<MelonPult> = registerPlant(
         "melonpult",
         EntityType.Builder.of(::MelonPult, MobCategory.CREATURE),
-        sunCost = 10,
         width = 0.9f,
         height = 0.8f,
         attributes = Plant.Companion.PlantAttributes(
@@ -178,7 +162,6 @@ object PazEntities {
     )
     @JvmField val PUFF_SHROOM: EntityType<PuffShroom> = registerPlant(
         "puffshroom", EntityType.Builder.of(::PuffShroom, MobCategory.CREATURE),
-        sunCost = 0,
         width = 0.5f,
         height = 0.65f,
         eyeHeight = 0.3f,
@@ -190,7 +173,6 @@ object PazEntities {
     )
     @JvmField val SCAREDY_SHROOM: EntityType<ScaredyShroom> = registerPlant(
         "scaredyshroom", EntityType.Builder.of(::ScaredyShroom, MobCategory.CREATURE),
-        sunCost = 3,
         width = 0.45f,
         height = 0.9f,
         eyeHeight = 0.5f,
@@ -202,7 +184,6 @@ object PazEntities {
     @JvmField val FUME_SHROOM: EntityType<FumeShroom> = registerPlant(
         "fumeshroom",
         EntityType.Builder.of(::FumeShroom, MobCategory.CREATURE),
-        sunCost = 6,
         width = 0.8f,
         height = 0.8f,
         attributes = Plant.Companion.PlantAttributes(
@@ -212,12 +193,10 @@ object PazEntities {
     @JvmField val SUN_SHROOM: EntityType<SunShroom> = registerPlant(
         "sunshroom",
         EntityType.Builder.of(::SunShroom, MobCategory.CREATURE),
-        sunCost = 3,
         height = 0.85f
     )
     @JvmField val HYPNOSHROOM: EntityType<HypnoShroom> = registerPlant(
         "hypnoshroom", EntityType.Builder.of(::HypnoShroom, MobCategory.CREATURE),
-        sunCost = 7,
         width = 0.6f,
         height = 1.3f,
         eyeHeight = 0.6f,
@@ -228,7 +207,6 @@ object PazEntities {
     )
     @JvmField val DOOM_SHROOM: EntityType<DoomShroom> = registerPlant(
         "doomshroom", EntityType.Builder.of(::DoomShroom, MobCategory.CREATURE),
-        sunCost = 12,
         eyeHeight = 0.6f,
         height = 0.8f,
         attributes = Plant.Companion.PlantAttributes(
@@ -238,7 +216,6 @@ object PazEntities {
     )
     @JvmField val COFFEE_BEAN: EntityType<CoffeeBean> = registerPlant(
         "coffeebean", EntityType.Builder.of(::CoffeeBean, MobCategory.CREATURE),
-        sunCost = 3,
         width = 0.4f,
         height = 0.5f,
         attributes = Plant.Companion.PlantAttributes(
@@ -433,7 +410,6 @@ object PazEntities {
     private fun <T : LivingEntity> registerPlant(
         name : String,
         builder: EntityType.Builder<T> = EntityType.Builder.createNothing(MobCategory.CREATURE),
-        sunCost: Int = 0,
         width: Float = 0.6f,
         height: Float = 1.0f,
         eyeHeight: Float = height * 0.85f,
@@ -441,8 +417,7 @@ object PazEntities {
     ): EntityType<T> {
         builder.sized(width, height).eyeHeight(eyeHeight)
         val type = register(name, builder)
-        FabricDefaultAttributeRegistry.register(type, attributes.copy(sunCost = sunCost).apply(createMobAttributes()))
-        typeToSunCost[type] = sunCost
+        FabricDefaultAttributeRegistry.register(type, attributes.apply(createMobAttributes()))
         return type
     }